Implement Perceptron algorithm - first attempt
This attempt includes a simple linearly seperable example which trains on input
data and plots the decision boundary between two classes.

This attempt is not optimized or vectorized. It also does not make very good use
of the python standard library, numpy, or scikit_learn. Future attempts will do
so.

This is meant as a baseline when comparing to more optimized implementations.